Move 'before_request' into 'main' blueprint
Update comment in api_handler with new before_request location
